removed old orchestrator code and moved around interfaces and typedefs
