--- nl.cwi.dis.vr2gather/Runtime/VRTOrchestrator/API/OrchestratorWSManagement.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e22d1e956528b114286f035921a09c89
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
