--- nl.cwi.dis.vr2gather/Runtime/VRTOrchestrator/API/OrchestratorWSManagement/OrchestratorCommandsAndResponses.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4188152a21ed0bb499580cad38f30d2d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
