--- nl.cwi.dis.vr2gather/Runtime/VRTOrchestrator/API/OrchestratorWSManagement/OrchestratorWSManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ff110c3155a729a4590aab8d6647f411
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
