+++ nl.cwi.dis.vr2gather/Runtime/VRTOrchestrator/API/OrchestratorWrapping/OrchestratorInterfaces.cs
@@ -22,10 +22,25 @@
 //  other intellectual property.
 
 using System.Collections.Generic;
+using VRT.Orchestrator.Responses;
 
 //Interfaces to be implemented to supervise the orchestrator
 namespace VRT.Orchestrator.Wrapping
 {
+    public interface IOrchestratorConnectionListener
+    {
+        void OnSocketConnect();
+        void OnSocketConnecting();
+        void OnSocketDisconnect();
+        void OnSocketError(ResponseStatus message);
+    }
+
+    // interface to implement to be updated from messages exchanged on the socketio
+    public interface IMessagesListener
+    {
+        void OnOrchestratorResponse(int commandID, int status, string response);
+        void OnOrchestratorRequest(string request);
+    }
     // Interface to implement to listen the user messages emitted spontaneously
     // by the orchestrator
     public interface IUserMessagesListener
