--- nl.cwi.dis.vr2gather/Runtime/VRTOrchestrator/API/OrchestratorWrapping/OrchestratorWrapper.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2d64b7d0232e48149a3330234143132d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
